update: 内部IDを32bitに更新

+++ weaverail-model/src/id_issuer.rs
@@ -6,7 +6,7 @@ use crate::model::id::WeaverailId;
 
 #[derive(ts_rs::TS, Clone, PartialEq, Debug, Default, Serialize, Deserialize)]
 pub struct IdIssuer {
-    current: Cell<u16>,
+    current: Cell<u32>,
 }
 impl IdIssuer {
     pub fn new() -> Self {
